Added a separate component for easier development
