+++ src/screens/AddShiftScreen.js
@@ -5,7 +5,7 @@ import Navbar from "../components/Navbar";
 import styles from "../styles/AddShift";
 import { ShiftTimerContext } from "../context/ShiftTimerContext";
 import ShiftGroupDropDown  from "../components/ShiftGroupDropDown";
-import DateTimePicker from "@react-native-community/datetimepicker";
+import AddShiftManually from "../components/AddShiftManually";
 
 const RADIUS = 45;
 const CIRCUMFERENCE = 2 * Math.PI * RADIUS;
@@ -14,7 +14,6 @@ const AddShiftScreen = () => {
     const { setBreakDuration, breakDuration, saveShift, elapsedTime, running, paused, startShift, pauseShift, resumeShift, stopShift, setIsModalVisible, isModalVisible, openModal, formatTime, setShiftDescription, setShiftName, shiftName, shiftDescription } = useContext(ShiftTimerContext);
 
     const [isRecordMode, setIsRecordMode] = useState(true);
-    const [inputDropDownVisible, setInputDropDownVisible] = useState(false);
 
     //Shift and break duration states
     const [duration, setDuration ] = useState("");
@@ -44,92 +43,7 @@ useEffect(() => {
       outputRange: [CIRCUMFERENCE, 0],
     });
 
-    // State for managing the date and time pickers
-const [isStartDatePickerVisible, setIsStartDatePickerVisible] = useState(false);
-const [isStartTimePickerVisible, setIsStartTimePickerVisible] = useState(false);
-const [selectedStartDate, setSelectedStartDate] = useState(new Date());
 
-// Show and hide handlers for the date and time pickers
-const showStartDatePicker = () => setIsStartDatePickerVisible(true);
-const hideStartDatePicker = () => setIsStartDatePickerVisible(false);
-const showStartTimePicker = () => setIsStartTimePickerVisible(true);
-const hideStartTimePicker = () => setIsStartTimePickerVisible(false);
-
-// Handle date selection
-const handleStartDateChange = (event, date) => {
-    if (!date) {
-        hideStartDatePicker(); // Close the picker if dismissed
-        return;
-    }
-    setSelectedStartDate(date); // Update the selected date
-    hideStartDatePicker(); // Hide the date picker
-    showStartTimePicker(); // Show the time picker
-};
-
-// Handle time selection
-const handleStartTimeChange = (event, time) => {
-    if (!time) {
-        hideStartTimePicker(); // Close the picker if dismissed
-        return;
-    }
-    // Combine the selected date and time into a single Date object
-    const updatedDate = new Date(
-        selectedStartDate.getFullYear(),
-        selectedStartDate.getMonth(),
-        selectedStartDate.getDate(),
-        time.getHours(),
-        time.getMinutes()
-    );
-    setSelectedStartDate(updatedDate); // Update the selected start date with time
-    hideStartTimePicker(); // Hide the time picker
-};
-
-// Format the selected date and time for display
-const formatDateTime = (date) => {
-    if (!date) return "Valitse aika"; // Fallback text if no date is provided
-    const options = { day: "2-digit", month: "2-digit", year: "numeric", hour: "2-digit", minute: "2-digit" };
-    return date.toLocaleString("fi-FI", options); // Format for Finnish locale
-};
-
-    // State for managing the end date and time pickers
-const [isEndDatePickerVisible, setIsEndDatePickerVisible] = useState(false);
-const [isEndTimePickerVisible, setIsEndTimePickerVisible] = useState(false);
-const [selectedEndDate, setSelectedEndDate] = useState(new Date());
-
-// Show and hide handlers for the end date and time pickers
-const showEndDatePicker = () => setIsEndDatePickerVisible(true);
-const hideEndDatePicker = () => setIsEndDatePickerVisible(false);
-const showEndTimePicker = () => setIsEndTimePickerVisible(true);
-const hideEndTimePicker = () => setIsEndTimePickerVisible(false);
-
-// Handle end date selection
-const handleEndDateChange = (event, date) => {
-    if (!date) {
-        hideEndDatePicker(); // Close the picker if dismissed
-        return;
-    }
-    setSelectedEndDate(date); // Update the selected end date
-    hideEndDatePicker(); // Hide the date picker
-    showEndTimePicker(); // Show the time picker
-};
-
-// Handle end time selection
-const handleEndTimeChange = (event, time) => {
-    if (!time) {
-        hideEndTimePicker(); // Close the picker if dismissed
-        return;
-    }
-    // Combine the selected date and time into a single Date object
-    const updatedDate = new Date(
-        selectedEndDate.getFullYear(),
-        selectedEndDate.getMonth(),
-        selectedEndDate.getDate(),
-        time.getHours(),
-        time.getMinutes()
-    );
-    setSelectedEndDate(updatedDate); // Update the selected end date with time
-    hideEndTimePicker(); // Hide the time picker
-};
 
     return (
         <View style={styles.wrapper}>
@@ -216,100 +130,8 @@ const handleEndTimeChange = (event, time) => {
             ) : (
                 
                 <View style={styles.container}>
-                    
-                    {/*Input View! */}
-                    {/* Nimi, kuvaus, aloitusaika, lopetusaika, tauko, päivämäärä */}
-
-                    <View style={styles.shiftDataInputRow}>
-            <Text style={styles.shiftDataLabel}>Aloitusaika</Text>
-            <TouchableOpacity style={styles.inputbutton} onPress={showStartDatePicker}>
-                <Text style={styles.buttonText}>{formatDateTime(selectedStartDate)}</Text>
-            </TouchableOpacity>
-        </View>
-
-        {/* Start Date Picker */}
-        {isStartDatePickerVisible && (
-            <DateTimePicker
-                value={selectedStartDate}
-                mode="date" // Show only the date picker
-                display="spinner"
-                onChange={handleStartDateChange}
-            />
-        )}
-
-        {/* Start Time Picker */}
-        {isStartTimePickerVisible && (
-            <DateTimePicker
-                value={selectedStartDate}
-                mode="time" // Show only the time picker
-                display="spinner"
-                onChange={handleStartTimeChange}
-            />
-        )}
-
-        {/* End Time Section */}
-        <View style={styles.shiftDataInputRow}>
-            <Text style={styles.shiftDataLabel}>Lopetusaika</Text>
-            <TouchableOpacity style={styles.inputbutton} onPress={showEndDatePicker}>
-                <Text style={styles.buttonText}>{formatDateTime(selectedEndDate)}</Text>
-            </TouchableOpacity>
-        </View>
-
-        {/* End Date Picker */}
-        {isEndDatePickerVisible && (
-            <DateTimePicker
-                value={selectedEndDate}
-                mode="date" // Show only the date picker
-                display="spinner"
-                onChange={handleEndDateChange}
-            />
-        )}
-
-        {/* End Time Picker */}
-        {isEndTimePickerVisible && (
-            <DateTimePicker
-                value={selectedEndDate}
-                mode="time" // Show only the time picker
-                display="spinner"
-                onChange={handleEndTimeChange}
-            />
-        )}
-
-        {/* Duration Field */}
-        <View style={styles.shiftDataInputRow}>
-            <Text style={styles.shiftDataLabel}>Kesto</Text>
-            <TextInput
-                style={styles.inputField}
-                placeholder="hh:mm"
-                value={duration}
-                onChangeText={setDuration}
-                keyboardType="numeric"
-            />
-        </View>
-
-            {/* Break Duration Field */}
-                <View style={styles.shiftDataInputRow}>
-                    <Text style={styles.shiftDataLabel}>Tauon kesto</Text>
-                    <TextInput
-                        style={styles.inputField}
-                        placeholder="hh:mm"
-                        value={breakDuration}
-                        onChangeText={setBreakDuration}
-                        keyboardType="numeric"
-                    />
+                    <AddShiftManually/>
                 </View>
-
-            {/* Save Button */}
-            <TouchableOpacity style={styles.button}
-            onPress={()=> {
-                console.log("Manual save triggered");
-                saveShift();
-            }}>
-                <Text style={styles.buttonText}>Tallenna</Text>
-            </TouchableOpacity>
-
-            </View>
-        
             )}
             </View>
         </View>
